+++ skcuda/cufft.py
@@ -12,19 +12,20 @@ import re
 import sys
 
 # Load library:
-_version_list = [10.0, 9.2, 9.1, 9.0, 8.0, 7.5, 7.0, 6.5, 6.0, 5.5, 5.0, 4.0]
+_linux_version_list = [10.1, 10.0, 9.2, 9.1, 9.0, 8.0, 7.5, 7.0, 6.5, 6.0, 5.5, 5.0, 4.0]
+_win32_version_list = [10, 100, 92, 91, 90, 80, 75, 70, 65, 60, 55, 50, 40]
 if 'linux' in sys.platform:
     _libcufft_libname_list = ['libcufft.so'] + \
-                             ['libcufft.so.%s' % v for v in _version_list]
+                             ['libcufft.so.%s' % v for v in _linux_version_list]
 elif sys.platform == 'darwin':
     _libcufft_libname_list = ['libcufft.dylib']
 elif sys.platform == 'win32':
     if sys.maxsize > 2**32:
         _libcufft_libname_list = ['cufft.dll'] + \
-                                 ['cufft64_%s.dll' % int(10*v) for v in _version_list]
+            ['cufft64_%s.dll' % v for v in _win32_version_list]
     else:
         _libcufft_libname_list = ['cufft.dll'] + \
-                                 ['cufft32_%s.dll' % int(10*v) for v in _version_list]
+            ['cufft32_%s.dll' % v for v in _win32_version_list]
 
 else:
     raise RuntimeError('unsupported platform')