+++ tests/test_cublas.py
@@ -640,7 +640,7 @@ class test_cublas(TestCase):
         from scipy.linalg import lu_factor
         l, m = 11, 7
         A = np.random.rand(l, m, m).astype(np.float32)
-        A = np.array([np.matrix(a)*np.matrix(a).T for a in A])
+        A = np.array([np.dot(a, a.T) for a in A])
 
         a_gpu = gpuarray.to_gpu(A)
         a_arr = bptrs(a_gpu)
@@ -660,7 +660,7 @@ class test_cublas(TestCase):
         from scipy.linalg import lu_factor
         l, m = 11, 7
         A = np.random.rand(l, m, m).astype(np.float64)
-        A = np.array([np.matrix(a)*np.matrix(a).T for a in A])
+        A = np.array([np.dot(a, a.T) for a in A])
 
         a_gpu = gpuarray.to_gpu(A)
         a_arr = bptrs(a_gpu)