+++ tests/test_fft.py
@@ -45,7 +45,7 @@ class test_fft(TestCase):
         xf_gpu = gpuarray.empty(self.N//2+1, np.complex64)
         plan = fft.Plan(x.shape, np.float32, np.complex64)
         fft.fft(x_gpu, xf_gpu, plan)
-        assert np.allclose(xf, xf_gpu.get(), atol=atol_float32)
+        np.testing.assert_allclose(xf, xf_gpu.get(), atol=atol_float32)
 
     def test_fft_float32_to_complex64_2d(self):
         x = np.asarray(np.random.rand(self.N, self.M), np.float32)
@@ -54,7 +54,7 @@ class test_fft(TestCase):
         xf_gpu = gpuarray.empty((self.N, self.M//2+1), np.complex64)
         plan = fft.Plan(x.shape, np.float32, np.complex64)
         fft.fft(x_gpu, xf_gpu, plan)
-        assert np.allclose(xf, xf_gpu.get(), atol=atol_float32)
+        np.testing.assert_allclose(xf, xf_gpu.get(), atol=atol_float32)
 
     def test_batch_fft_float32_to_complex64_1d(self):
         x = np.asarray(np.random.rand(self.B, self.N), np.float32)
@@ -63,7 +63,7 @@ class test_fft(TestCase):
         xf_gpu = gpuarray.empty((self.B, self.N//2+1), np.complex64)
         plan = fft.Plan(x.shape[1], np.float32, np.complex64, batch=self.B)
         fft.fft(x_gpu, xf_gpu, plan)
-        assert np.allclose(xf, xf_gpu.get(), atol=atol_float32)
+        np.testing.assert_allclose(xf, xf_gpu.get(), atol=atol_float32)
 
     def test_batch_fft_float32_to_complex64_2d(self):
         x = np.asarray(np.random.rand(self.B, self.N, self.M), np.float32)
@@ -72,7 +72,7 @@ class test_fft(TestCase):
         xf_gpu = gpuarray.empty((self.B, self.N, self.M//2+1), np.complex64)
         plan = fft.Plan([self.N, self.M], np.float32, np.complex64, batch=self.B)
         fft.fft(x_gpu, xf_gpu, plan)
-        assert np.allclose(xf, xf_gpu.get(), atol=atol_float32)
+        np.testing.assert_allclose(xf, xf_gpu.get(), atol=atol_float32)
 
     def test_fft_float64_to_complex128_1d(self):
         x = np.asarray(np.random.rand(self.N), np.float64)
@@ -81,7 +81,7 @@ class test_fft(TestCase):
         xf_gpu = gpuarray.empty(self.N//2+1, np.complex128)
         plan = fft.Plan(x.shape, np.float64, np.complex128)
         fft.fft(x_gpu, xf_gpu, plan)
-        assert np.allclose(xf, xf_gpu.get(), atol=atol_float64)
+        np.testing.assert_allclose(xf, xf_gpu.get(), atol=atol_float64)
 
     def test_fft_float64_to_complex128_2d(self):
         x = np.asarray(np.random.rand(self.N, self.M), np.float64)
@@ -90,7 +90,7 @@ class test_fft(TestCase):
         xf_gpu = gpuarray.empty((self.N, self.M//2+1), np.complex128)
         plan = fft.Plan(x.shape, np.float64, np.complex128)
         fft.fft(x_gpu, xf_gpu, plan)
-        assert np.allclose(xf, xf_gpu.get(), atol=atol_float64)
+        np.testing.assert_allclose(xf, xf_gpu.get(), atol=atol_float64)
 
     def test_batch_fft_float64_to_complex128_1d(self):
         x = np.asarray(np.random.rand(self.B, self.N), np.float64)
@@ -99,7 +99,7 @@ class test_fft(TestCase):
         xf_gpu = gpuarray.empty((self.B, self.N//2+1), np.complex128)
         plan = fft.Plan(x.shape[1], np.float64, np.complex128, batch=self.B)
         fft.fft(x_gpu, xf_gpu, plan)
-        assert np.allclose(xf, xf_gpu.get(), atol=atol_float64)
+        np.testing.assert_allclose(xf, xf_gpu.get(), atol=atol_float64)
 
     def test_batch_fft_float64_to_complex128_2d(self):
         x = np.asarray(np.random.rand(self.B, self.N, self.M), np.float64)
@@ -108,7 +108,7 @@ class test_fft(TestCase):
         xf_gpu = gpuarray.empty((self.B, self.N, self.M//2+1), np.complex128)
         plan = fft.Plan([self.N, self.M], np.float64, np.complex128, batch=self.B)
         fft.fft(x_gpu, xf_gpu, plan)
-        assert np.allclose(xf, xf_gpu.get(), atol=atol_float64)
+        np.testing.assert_allclose(xf, xf_gpu.get(), atol=atol_float64)
 
     def test_ifft_complex64_to_float32_1d(self):
         x = np.asarray(np.random.rand(self.N), np.float32)
@@ -117,7 +117,7 @@ class test_fft(TestCase):
         x_gpu = gpuarray.empty(self.N, np.float32)
         plan = fft.Plan(x.shape, np.complex64, np.float32)
         fft.ifft(xf_gpu, x_gpu, plan, True)
-        assert np.allclose(x, x_gpu.get(), atol=atol_float32)
+        np.testing.assert_allclose(x, x_gpu.get(), atol=atol_float32)
 
     def test_ifft_complex64_to_float32_2d(self):
 
@@ -130,7 +130,7 @@ class test_fft(TestCase):
         x_gpu = gpuarray.empty((self.N, self.M), np.float32)
         plan = fft.Plan(x.shape, np.complex64, np.float32)
         fft.ifft(xf_gpu, x_gpu, plan, True)
-        assert np.allclose(x, x_gpu.get(), atol=atol_float32)
+        np.testing.assert_allclose(x, x_gpu.get(), atol=atol_float32)
 
     def test_batch_ifft_complex64_to_float32_1d(self):
 
@@ -143,7 +143,7 @@ class test_fft(TestCase):
         x_gpu = gpuarray.empty((self.B, self.N), np.float32)
         plan = fft.Plan(x.shape[1], np.complex64, np.float32, batch=self.B)
         fft.ifft(xf_gpu, x_gpu, plan, True)
-        assert np.allclose(x, x_gpu.get(), atol=atol_float32)
+        np.testing.assert_allclose(x, x_gpu.get(), atol=atol_float32)
 
     def test_batch_ifft_complex64_to_float32_2d(self):
 
@@ -156,7 +156,7 @@ class test_fft(TestCase):
         x_gpu = gpuarray.empty((self.B, self.N, self.M), np.float32)
         plan = fft.Plan([self.N, self.M], np.complex64, np.float32, batch=self.B)
         fft.ifft(xf_gpu, x_gpu, plan, True)
-        assert np.allclose(x, x_gpu.get(), atol=atol_float32)
+        np.testing.assert_allclose(x, x_gpu.get(), atol=atol_float32)
 
     def test_ifft_complex128_to_float64_1d(self):
         x = np.asarray(np.random.rand(self.N), np.float64)
@@ -165,7 +165,7 @@ class test_fft(TestCase):
         x_gpu = gpuarray.empty(self.N, np.float64)
         plan = fft.Plan(x.shape, np.complex128, np.float64)
         fft.ifft(xf_gpu, x_gpu, plan, True)
-        assert np.allclose(x, x_gpu.get(), atol=atol_float64)
+        np.testing.assert_allclose(x, x_gpu.get(), atol=atol_float64)
 
     def test_ifft_complex128_to_float64_2d(self):
 
@@ -178,7 +178,7 @@ class test_fft(TestCase):
         x_gpu = gpuarray.empty((self.N, self.M), np.float64)
         plan = fft.Plan(x.shape, np.complex128, np.float64)
         fft.ifft(xf_gpu, x_gpu, plan, True)
-        assert np.allclose(x, x_gpu.get(), atol=atol_float64)
+        np.testing.assert_allclose(x, x_gpu.get(), atol=atol_float64)
 
     def test_batch_ifft_complex128_to_float64_1d(self):
 
@@ -191,7 +191,7 @@ class test_fft(TestCase):
         x_gpu = gpuarray.empty((self.B, self.N), np.float64)
         plan = fft.Plan(x.shape[1], np.complex128, np.float64, batch=self.B)
         fft.ifft(xf_gpu, x_gpu, plan, True)
-        assert np.allclose(x, x_gpu.get(), atol=atol_float64)
+        np.testing.assert_allclose(x, x_gpu.get(), atol=atol_float64)
 
     def test_batch_ifft_complex128_to_float64_2d(self):
 
@@ -204,7 +204,7 @@ class test_fft(TestCase):
         x_gpu = gpuarray.empty((self.B, self.N, self.M), np.float64)
         plan = fft.Plan([self.N, self.M], np.complex128, np.float64, batch=self.B)
         fft.ifft(xf_gpu, x_gpu, plan, True)
-        assert np.allclose(x, x_gpu.get(), atol=atol_float64)
+        np.testing.assert_allclose(x, x_gpu.get(), atol=atol_float64)
 
     def test_multiple_streams(self):
         x = np.asarray(np.random.rand(self.N), np.float32)
@@ -221,8 +221,8 @@ class test_fft(TestCase):
         plan2 = fft.Plan(y.shape, np.float32, np.complex64, stream=stream1)
         fft.fft(x_gpu, xf_gpu, plan1)
         fft.fft(y_gpu, yf_gpu, plan2)
-        assert np.allclose(xf, xf_gpu.get(), atol=atol_float32)
-        assert np.allclose(yf, yf_gpu.get(), atol=atol_float32)
+        np.testing.assert_allclose(xf, xf_gpu.get(), atol=atol_float32)
+        np.testing.assert_allclose(yf, yf_gpu.get(), atol=atol_float32)
 
     def test_work_area(self):
         x = np.asarray(np.random.rand(self.N), np.float32)
@@ -233,7 +233,7 @@ class test_fft(TestCase):
         work_area = gpuarray.empty((plan.worksize,), np.uint8)
         plan.set_work_area(work_area)
         fft.fft(x_gpu, xf_gpu, plan)
-        assert np.allclose(xf, xf_gpu.get(), atol=atol_float32)
+        np.testing.assert_allclose(xf, xf_gpu.get(), atol=atol_float32)
 
 def suite():
     context = make_default_context()