+++ tests/test_integrate.py
@@ -33,89 +33,89 @@ class test_integrate(TestCase):
         x = np.asarray(np.random.rand(10), np.float32)
         x_gpu = gpuarray.to_gpu(x)
         z = integrate.trapz(x_gpu)
-        assert np.allclose(np.trapz(x), z)
+        np.testing.assert_allclose(np.trapz(x), z)
 
     def test_trapz_float64(self):
         x = np.asarray(np.random.rand(10), np.float64)
         x_gpu = gpuarray.to_gpu(x)
         z = integrate.trapz(x_gpu)
-        assert np.allclose(np.trapz(x), z)
+        np.testing.assert_allclose(np.trapz(x), z)
 
     def test_trapz_complex64(self):
         x = np.asarray(np.random.rand(10)+1j*np.random.rand(10), np.complex64)
         x_gpu = gpuarray.to_gpu(x)
         z = integrate.trapz(x_gpu)
-        assert np.allclose(np.trapz(x), z)
+        np.testing.assert_allclose(np.trapz(x), z)
 
     def test_trapz_complex128(self):
         x = np.asarray(np.random.rand(10)+1j*np.random.rand(10), np.complex128)
         x_gpu = gpuarray.to_gpu(x)
         z = integrate.trapz(x_gpu)
-        assert np.allclose(np.trapz(x), z)
+        np.testing.assert_allclose(np.trapz(x), z)
 
     def test_simps_float32(self):
         x = np.asarray(np.random.rand(10), np.float32)
         x_gpu = gpuarray.to_gpu(x)
-        assert np.allclose(scipy.integrate.simps(x), 
+        np.testing.assert_allclose(scipy.integrate.simps(x),
                         integrate.simps(x_gpu))
-        assert np.allclose(scipy.integrate.simps(x, even='first'), 
+        np.testing.assert_allclose(scipy.integrate.simps(x, even='first'),
                         integrate.simps(x_gpu, even='first'))
-        assert np.allclose(scipy.integrate.simps(x, even='last'), 
+        np.testing.assert_allclose(scipy.integrate.simps(x, even='last'),
                         integrate.simps(x_gpu, even='last'))
 
     def test_simps_float64(self):
         x = np.asarray(np.random.rand(10), np.float64)
         x_gpu = gpuarray.to_gpu(x)
-        assert np.allclose(scipy.integrate.simps(x), 
+        np.testing.assert_allclose(scipy.integrate.simps(x),
                         integrate.simps(x_gpu))
-        assert np.allclose(scipy.integrate.simps(x, even='first'), 
+        np.testing.assert_allclose(scipy.integrate.simps(x, even='first'),
                         integrate.simps(x_gpu, even='first'))
-        assert np.allclose(scipy.integrate.simps(x, even='last'), 
+        np.testing.assert_allclose(scipy.integrate.simps(x, even='last'),
                         integrate.simps(x_gpu, even='last'))
 
     def test_simps_complex64(self):
         x = np.asarray(np.random.rand(10)+1j*np.random.rand(10), np.complex64)
         x_gpu = gpuarray.to_gpu(x)
-        assert np.allclose(scipy.integrate.simps(x), 
+        np.testing.assert_allclose(scipy.integrate.simps(x),
                         integrate.simps(x_gpu))
-        assert np.allclose(scipy.integrate.simps(x, even='first'), 
+        np.testing.assert_allclose(scipy.integrate.simps(x, even='first'),
                         integrate.simps(x_gpu, even='first'))
-        assert np.allclose(scipy.integrate.simps(x, even='last'), 
+        np.testing.assert_allclose(scipy.integrate.simps(x, even='last'),
                         integrate.simps(x_gpu, even='last'))
 
     def test_simps_complex128(self):
         x = np.asarray(np.random.rand(10)+1j*np.random.rand(10), np.complex128)
         x_gpu = gpuarray.to_gpu(x)
-        assert np.allclose(scipy.integrate.simps(x), 
+        np.testing.assert_allclose(scipy.integrate.simps(x),
                         integrate.simps(x_gpu))
-        assert np.allclose(scipy.integrate.simps(x, even='first'), 
+        np.testing.assert_allclose(scipy.integrate.simps(x, even='first'),
                         integrate.simps(x_gpu, even='first'))
-        assert np.allclose(scipy.integrate.simps(x, even='last'), 
+        np.testing.assert_allclose(scipy.integrate.simps(x, even='last'),
                         integrate.simps(x_gpu, even='last'))
 
     def test_trapz2d_float32(self):
         x = np.asarray(np.random.rand(5, 5), np.float32)
         x_gpu = gpuarray.to_gpu(x)
         z = integrate.trapz2d(x_gpu)
-        assert np.allclose(np.trapz(np.trapz(x)), z)
+        np.testing.assert_allclose(np.trapz(np.trapz(x)), z)
 
     def test_trapz2d_float64(self):
         x = np.asarray(np.random.rand(5, 5), np.float64)
         x_gpu = gpuarray.to_gpu(x)
         z = integrate.trapz2d(x_gpu)
-        assert np.allclose(np.trapz(np.trapz(x)), z)
+        np.testing.assert_allclose(np.trapz(np.trapz(x)), z)
 
     def test_trapz2d_complex64(self):
         x = np.asarray(np.random.rand(5, 5)+1j*np.random.rand(5, 5), np.complex64)
         x_gpu = gpuarray.to_gpu(x)
         z = integrate.trapz2d(x_gpu)
-        assert np.allclose(np.trapz(np.trapz(x)), z)
+        np.testing.assert_allclose(np.trapz(np.trapz(x)), z)
 
     def test_trapz2d_complex128(self):
         x = np.asarray(np.random.rand(5, 5)+1j*np.random.rand(5, 5), np.complex128)
         x_gpu = gpuarray.to_gpu(x)
         z = integrate.trapz2d(x_gpu)
-        assert np.allclose(np.trapz(np.trapz(x)), z)
+        np.testing.assert_allclose(np.trapz(np.trapz(x)), z)
 
 def suite():
     context = make_default_context()